requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
langchain-pinecone
sentence-transformers
pinecone
chromadb
flask
python-dotenv
PyPDF2
python-dotenv
-e .